# backend/internal/pkg/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HRIS Platform API",
...
    "version": "1.5.0",
    "contact": {
      "name": "HRIS Platform Team",
      "email": "dev@hris-platform.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development"
    }
  ],
...
  "paths": {
...
    "/api/v1/tenant/competency/competencies": {
      "get": {
        "tags": ["Tenant: Competency Management"],
        "summary": "List competencies with pagination",
        "operationId": "listCompetencies",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Paginated list of competencies" } }
      },
      "post": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Create a new competency master",
        "operationId": "createCompetency",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCompetencyRequest" }
            }
          }
        },
        "responses": { "201": { "description": "Competency created" }, "400": { "description": "Validation error" } }
      }
    },
    "/api/v1/tenant/competency/competencies/{id}": {
      "get": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Get competency by ID",
        "operationId": "getCompetency",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Competency details" }, "404": { "description": "Not found" } }
      },
      "put": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Update competency",
        "operationId": "updateCompetency",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCompetencyRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Competency updated" } }
      },
      "delete": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Delete competency",
        "operationId": "deleteCompetency",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Competency deleted" } }
      }
    },
    "/api/v1/tenant/competency/competence-values": {
      "get": {
        "tags": ["Tenant: Competency Management"],
        "summary": "List competence values (legacy) with pagination",
        "operationId": "listCompetenceValues",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Paginated list of competence values" } }
      },
      "post": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Create a new competence value (legacy)",
        "operationId": "createCompetenceValue",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCompetenceValueRequest" }
            }
          }
        },
        "responses": { "201": { "description": "Competence value created" }, "400": { "description": "Validation error" } }
      }
    },
    "/api/v1/tenant/competency/competence-values/{id}": {
      "get": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Get competence value by ID",
        "operationId": "getCompetenceValue",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Competence value details" }, "404": { "description": "Not found" } }
      },
      "put": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Update competence value",
        "operationId": "updateCompetenceValue",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCompetenceValueRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Competence value updated" } }
      },
      "delete": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Delete competence value",
        "operationId": "deleteCompetenceValue",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Competence value deleted" } }
      }
    },
    "/api/v1/tenant/competency/values": {
      "get": {
        "tags": ["Tenant: Competency Management"],
        "summary": "List competency values (structured) with pagination",
        "operationId": "listCompetencyValues",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Paginated list of competency values" } }
      },
      "post": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Create a new competency value (structured)",
        "operationId": "createCompetencyValue",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCompetencyValueRequest" }
            }
          }
        },
        "responses": { "201": { "description": "Competency value created" }, "400": { "description": "Validation error" } }
      }
    },
    "/api/v1/tenant/competency/values/{id}": {
      "get": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Get competency value by ID",
        "operationId": "getCompetencyValue",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Competency value details" }, "404": { "description": "Not found" } }
      },
      "put": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Update competency value",
        "operationId": "updateCompetencyValue",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCompetencyValueRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Competency value updated" } }
      },
      "delete": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Delete competency value",
        "operationId": "deleteCompetencyValue",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Competency value deleted" } }
      }
    },
    "/api/v1/tenant/competency/events": {
      "get": {
        "tags": ["Tenant: Competency Management"],
        "summary": "List competency events with pagination",
        "operationId": "listCompetencyEvents",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Paginated list of competency events" } }
      },
      "post": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Create a new competency event",
        "operationId": "createCompetencyEvent",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCompetencyEventRequest" }
            }
          }
        },
        "responses": { "201": { "description": "Competency event created" }, "400": { "description": "Validation error" } }
      }
    },
    "/api/v1/tenant/competency/events/{id}": {
      "get": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Get competency event by ID",
        "operationId": "getCompetencyEvent",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Competency event details" }, "404": { "description": "Not found" } }
      },
      "put": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Update competency event",
        "operationId": "updateCompetencyEvent",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCompetencyEventRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Competency event updated" } }
      },
      "delete": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Delete competency event",
        "operationId": "deleteCompetencyEvent",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Competency event deleted" } }
      }
    },
    "/api/v1/tenant/competency/event-targets": {
      "get": {
        "tags": ["Tenant: Competency Management"],
        "summary": "List competency event targets with pagination",
        "operationId": "listCompetencyEventTargets",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Paginated list of competency event targets" } }
      },
      "post": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Create a new competency event target",
        "operationId": "createCompetencyEventTarget",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCompetencyEventTargetRequest" }
            }
          }
        },
        "responses": { "201": { "description": "Competency event target created" }, "400": { "description": "Validation error" } }
      }
    },
    "/api/v1/tenant/competency/event-targets/{id}": {
      "get": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Get competency event target by ID",
        "operationId": "getCompetencyEventTarget",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Competency event target details" }, "404": { "description": "Not found" } }
      },
      "put": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Update competency event target",
        "operationId": "updateCompetencyEventTarget",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCompetencyEventTargetRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Competency event target updated" } }
      },
      "delete": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Delete competency event target",
        "operationId": "deleteCompetencyEventTarget",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Competency event target deleted" } }
      }
    },
    "/api/v1/tenant/competency/scores": {
      "get": {
        "tags": ["Tenant: Competency Management"],
        "summary": "List competency scores with pagination",
        "operationId": "listCompetencyScores",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Paginated list of competency scores" } }
      },
      "post": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Create a new competency score",
        "operationId": "createCompetencyScore",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCompetencyScoreRequest" }
            }
          }
        },
        "responses": { "201": { "description": "Competency score created" }, "400": { "description": "Validation error" } }
      }
    },
    "/api/v1/tenant/competency/scores/{id}": {
      "get": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Get competency score by ID with details",
        "operationId": "getCompetencyScore",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Competency score details with score details" }, "404": { "description": "Not found" } }
      },
      "put": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Update competency score",
        "operationId": "updateCompetencyScore",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCompetencyScoreRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Competency score updated" } }
      },
      "delete": {
        "tags": ["Tenant: Competency Management"],
        "summary": "Delete competency score",
        "operationId": "deleteCompetencyScore",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Competency score deleted" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Access Token. Dapatkan dengan login ke platform endpoint."
      }
    },
    "schemas": {
...
      "CreateCompetencyRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "maxLength": 255, "example": "Leadership" },
          "field": { "type": "string", "maxLength": 255 },
          "cluster": { "type": "string", "maxLength": 255 },
          "definition": { "type": "string" }
        }
      },
      "UpdateCompetencyRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 255, "nullable": true },
          "field": { "type": "string", "maxLength": 255, "nullable": true },
          "cluster": { "type": "string", "maxLength": 255, "nullable": true },
          "definition": { "type": "string", "nullable": true }
        }
      },
...
      "CreateCompetenceValueRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "type": { "type": "string", "maxLength": 255 },
          "level": { "type": "integer" },
          "name": { "type": "string", "maxLength": 255, "example": "Excellent" },
          "point": { "type": "integer" },
          "description": { "type": "string", "maxLength": 255 }
        }
      },
      "UpdateCompetenceValueRequest": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "maxLength": 255, "nullable": true },
          "level": { "type": "integer", "nullable": true },
          "name": { "type": "string", "maxLength": 255, "nullable": true },
          "point": { "type": "integer", "nullable": true },
          "description": { "type": "string", "maxLength": 255, "nullable": true }
        }
      },
...
    }
  }
}